requirements.txt
absl-py
asgiref
astor
astunparse
attrs
backcall
cachetools
certifi
chardet
click
colorama
cycler
decorator
defusedxml
Django
easydict
entrypoints
enum34
Flask
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
image
imageio
implicits
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-console
jupyter-core
kafka-python
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
nbconvert
nbformat
networkx
notebook
numpy
oauthlib
opencv-python
opt-einsum
pandocfilters
parso
pexpect
pickleshare
Pillow
pip-date
portpicker
prometheus-client
prompt-toolkit
protobuf
ptyprocess
py4j
pyasn1
pyasn1-modules
pycocotools
pydot
Pygments
pyparsing
pyrsistent
pyspark
pytesseract
python-dateutil
pytz
PyWavelets
pywin32
pywinpty
PyYAML
pyzmq
qtconsole
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
Send2Trash
six
sklearn
sqlparse
tensorboard
tensorboard-plugin-wit
tensorflow-gpu-estimator
termcolor
terminado
testpath
tf-utils
traitlets
urllib3
wcwidth
Werkzeug
widgetsnbextension
wincertstore
wrapt
zipp
